Move functions around.
