+++ internal/service/s3/object.go
@@ -641,181 +641,6 @@ func findObject(ctx context.Context, conn *s3.Client, input *s3.HeadObjectInput)
 	return output, nil
 }
 
-// deleteAllObjectVersions deletes all versions of a specified key from an S3 bucket.
-// If key is empty then all versions of all objects are deleted.
-// Set force to true to override any S3 object lock protections on object lock enabled buckets.
-// Returns the number of objects deleted.
-func deleteAllObjectVersions(ctx context.Context, conn *s3.Client, bucketName, key string, force, ignoreObjectErrors bool) (int64, error) {
-	var nObjects int64
-
-	input := &s3.ListObjectVersionsInput{
-		Bucket: aws.String(bucketName),
-	}
-	if key != "" {
-		input.Prefix = aws.String(key)
-	}
-	var lastErr error
-
-	pages := s3.NewListObjectVersionsPaginator(conn, input)
-	for pages.HasMorePages() {
-		page, err := pages.NextPage(ctx)
-
-		if tfawserr.ErrCodeEquals(err, errCodeNoSuchBucket) {
-			break
-		}
-
-		if err != nil {
-			return nObjects, err
-		}
-
-		for _, objectVersion := range page.Versions {
-			objectKey := aws.ToString(objectVersion.Key)
-			objectVersionID := aws.ToString(objectVersion.VersionId)
-
-			if key != "" && key != objectKey {
-				continue
-			}
-
-			err := deleteObjectVersion(ctx, conn, bucketName, objectKey, objectVersionID, force)
-
-			if err == nil {
-				nObjects++
-			}
-
-			if tfawserr.ErrCodeEquals(err, errCodeAccessDenied) && force {
-				// Remove any legal hold.
-				input := &s3.HeadObjectInput{
-					Bucket:    aws.String(bucketName),
-					Key:       aws.String(objectKey),
-					VersionId: aws.String(objectVersionID),
-				}
-
-				output, err := conn.HeadObject(ctx, input)
-
-				if err != nil {
-					log.Printf("[ERROR] Getting S3 Bucket (%s) Object (%s) Version (%s) metadata: %s", bucketName, objectKey, objectVersionID, err)
-					lastErr = err
-					continue
-				}
-
-				if output.ObjectLockLegalHoldStatus == types.ObjectLockLegalHoldStatusOn {
-					input := &s3.PutObjectLegalHoldInput{
-						Bucket: aws.String(bucketName),
-						Key:    aws.String(objectKey),
-						LegalHold: &types.ObjectLockLegalHold{
-							Status: types.ObjectLockLegalHoldStatusOff,
-						},
-						VersionId: aws.String(objectVersionID),
-					}
-
-					_, err := conn.PutObjectLegalHold(ctx, input)
-
-					if err != nil {
-						log.Printf("[ERROR] Putting S3 Bucket (%s) Object (%s) Version(%s) legal hold: %s", bucketName, objectKey, objectVersionID, err)
-						lastErr = err
-						continue
-					}
-
-					// Attempt to delete again.
-					err = deleteObjectVersion(ctx, conn, bucketName, objectKey, objectVersionID, force)
-
-					if err != nil {
-						lastErr = err
-					} else {
-						nObjects++
-					}
-
-					continue
-				}
-
-				// AccessDenied for another reason.
-				lastErr = fmt.Errorf("deleting S3 Bucket (%s) Object (%s) Version (%s): %w", bucketName, objectKey, objectVersionID, err)
-				continue
-			}
-
-			if err != nil {
-				lastErr = err
-			}
-		}
-	}
-
-	if lastErr != nil {
-		if !ignoreObjectErrors {
-			return nObjects, fmt.Errorf("deleting at least one S3 Object version, last error: %w", lastErr)
-		}
-
-		lastErr = nil
-	}
-
-	pages = s3.NewListObjectVersionsPaginator(conn, input)
-	for pages.HasMorePages() {
-		page, err := pages.NextPage(ctx)
-
-		if tfawserr.ErrCodeEquals(err, errCodeNoSuchBucket) {
-			break
-		}
-
-		if err != nil {
-			return nObjects, err
-		}
-
-		for _, deleteMarker := range page.DeleteMarkers {
-			deleteMarkerKey := aws.ToString(deleteMarker.Key)
-			deleteMarkerVersionID := aws.ToString(deleteMarker.VersionId)
-
-			if key != "" && key != deleteMarkerKey {
-				continue
-			}
-
-			// Delete markers have no object lock protections.
-			err := deleteObjectVersion(ctx, conn, bucketName, deleteMarkerKey, deleteMarkerVersionID, false)
-
-			if err != nil {
-				lastErr = err
-			} else {
-				nObjects++
-			}
-		}
-	}
-
-	if lastErr != nil {
-		if !ignoreObjectErrors {
-			return nObjects, fmt.Errorf("deleting at least one S3 Object delete marker, last error: %w", lastErr)
-		}
-	}
-
-	return nObjects, nil
-}
-
-// deleteObjectVersion deletes a specific object version.
-// Set force to true to override any S3 object lock protections.
-func deleteObjectVersion(ctx context.Context, conn *s3.Client, b, k, v string, force bool) error {
-	input := &s3.DeleteObjectInput{
-		Bucket: aws.String(b),
-		Key:    aws.String(k),
-	}
-
-	if v != "" {
-		input.VersionId = aws.String(v)
-	}
-	if force {
-		input.BypassGovernanceRetention = true
-	}
-
-	log.Printf("[INFO] Deleting S3 Bucket (%s) Object (%s) Version (%s)", b, k, v)
-	_, err := conn.DeleteObject(ctx, input)
-
-	if err != nil {
-		log.Printf("[WARN] Deleting S3 Bucket (%s) Object (%s) Version (%s): %s", b, k, v, err)
-	}
-
-	if tfawserr.ErrCodeEquals(err, errCodeNoSuchBucket, errCodeNoSuchKey) {
-		return nil
-	}
-
-	return err
-}
-
 func expandObjectDate(v string) *time.Time {
 	t, err := time.Parse(time.RFC3339, v)
 	if err != nil {
